+++ test/Export-DataJourney.Tests.ps1
@@ -14,7 +14,7 @@ Describe Export-DataJourney {
             $Layer | Add-ArchDataModel -Title mymodel -Class original
             $SubLayer = $Layer | Add-ArchDataLayer -Key mylayer -Title sublayer -PassThru
             $SubLayer | Add-ArchDataModel -Title mysubmodel -Class analysis
-            $Layer | Add-ArchDataFlow -Key mm -Title myflow -Source mymodel -Sink mysubmodel
+            $Layer | Add-ArchDataFlow -Key mm -Title myflow -Description mymodeldescription -Source mymodel -Sink mysubmodel
         }
 
         It works {
@@ -38,6 +38,7 @@ Describe Export-DataJourney {
             $imported.Layer[0].Flows.Count | Should -Be 1
             $imported.Layer[0].Flows[0].Key | Should -Be mm
             $imported.Layer[0].Flows[0].Title | Should -Be myflow
+            $imported.Layer[0].Flows[0].Description | Should -Be mymodeldescription
             $imported.Layer[0].Flows[0].Sources.Count | Should -Be 1
             $imported.Layer[0].Flows[0].Sources[0] | Should -Be mymodel
             $imported.Layer[0].Flows[0].Sinks.Count | Should -Be 1
@@ -84,6 +85,7 @@ Describe Export-DataJourney {
             $imported = Get-Content -Path $TestDrive\subfolder\layer\foo\flows\mm.yml -Raw | ConvertFrom-Yaml
             $imported.Key | Should -BeNullOrEmpty
             $imported.Title | Should -Be myflow
+            $imported.Description | Should -Be mymodeldescription
             $imported.Sources.Count | Should -Be 1
             $imported.Sources[0] | Should -Be mymodel
             $imported.Sinks.Count | Should -Be 1